fix undefined about require moudle `createMenu`
